Check whole plugin install message
It can happen, that plugin has dependecy on another plugin with similar name (e.g. git plugin has as a depedency git-client plugin). In this case when the dependecy is installed, log watcher can mis-recognize it as installation of plugin itself and can move on too soon, resulting in error. Checking for whole logger message shoudl prevent it.