sendKeys doesn't always work.
sendKeys really just simulate sending keys.
Browsers try to select the option that matches the typed text,
but there are some keys that are interpreted differently, such
as SPACE, which selects the next option from the current option.

A more reliably way to select a specific option is to call
click on the <option> element.

This allows selection by not only the text but also the value.