wait until we find a visible element.
if none of the element is visible, rather than returning
with the first element and hope for the best, wait until
something becomes visible.

this is useful because Jenkins often changes visibility of
elements in reaction to the user action, and that can take
a bit of time.

So if the find method runs too fast, it can end up picking
up a non-visible element.