Redoing the way we obtain form-element-path.hpi
How we obtain this test assistance plugin is almost always orthogonal to
how we want to launch Jenkins, and the former is mostly uninteresting to
users but the latter is of great interest to users, so it makes sense to
separate them.

It also keeps the wiring script free from implementation details.