Tweaking the relative path resolution rules.
Sometimes we come across the path expression that ends with "/" like "/attributes[1]/" --- those are select controls
that produces different config fragments depending on the option selected.

We want to be able to create such references via control("/").
Note that control("") is already used throughout for referring to a path without "/" in the end, so we need to keep that working