Performance improvement.
On Firefox, driver.switchTo().alert() takes about 1sec. This callback gets called every time
a checkbox/radio button or something like that is clicked, this adds a significant delay.

So adding a fast-path check to leave this call quickly.

Personally, I'd question the wisdom of doing this check on every clicking. It seems to me
that doing this check only on a page navigation is sufficient.