Add functionality to log ProcessBuilder issues.
Add a setup method which initializes the logger before the test.

Add a logProcessBuilderIssues() method which detects whether or not
a process started by a ProcessBuilder has completed successfully
(return value of 0), or not. Any unsuccessful attempts are logged.

Add a logCurlHttpCodeIssues() method which checks the http
response code returned when running a curl command. Any
unsuccessful status codes such as 401, 404, etc. are logged.