Create an Exercised Plugins Report
This feature permits the creation of a report that describes
which plugins were exercised and what versions were used.

Two reporter types are available. The default "Console" reporter and
the "Text File" reporter.

The Text File reporter creates a java properties file and is useful
when wanting to be able to see which plugins and their versions were
tested with a particular version of Jenkins Core.

[FIXED JENKINS-23754]