JUT server was broken
The code was assuming 'jenkins_updates' is available as a directory, which is true only when these tests are run locally with "mvn test".
But it is not the case when ./jut-server is used, or when tests that reside in another module is executed with test harness compiled as a jar file.

This change moves these templates into a zip file to avoid this problem