[JENKINS-24436]
Fixing a test failure.

https://jenkins.ci.cloudbees.com/job/core/job/acceptance-test-harness-stable/361/ reports 8 consective failures on this test case because of the following assertion error:

    Expected: is "job/archaeological_replacement/checkstyle"
         but: was "/view/difficult_style/job/archaeological_replacement/checkstyle"

IIUC, the spirit of the test (and the bug reported in JENKINS-24436) is to make sure that the link correctly takes the user to the report page. I think the test would be more robust and sufficient if we just check that, instead of insisting on a relative url vs path-absolute URL.