Wrong hard-coding of numbers.
Jenkins with different plugins installed will not have the same list of SecurityRealms and AuthorizationStrategies in this order.
Better to reuse the logic implemented in GlobalSecurityConfig.

In general I don't think it's wise to rebuild parallel page object tree just for Geb. Instead, we should figure out a way
to bridge existing Java Page objects into Geb, for example by having Geb PO delegate to Jenkins PO. They all build on the
same WebDriver, so this should be quite possible.