[JENKINS-33520] DumbSlaves should launch on windows.
DumbSlave launching was not working correctly on windows even if cygwin
with curl was on the path for the invoked maven session.

The new process (for whatever reason) was not created as a child of the
spawning process and did not inherit its environment and so was inheriting
the default envirnonment.  WOrst than that is that the slave root was set
to /tmp which means you may get lucky and be able to create a folder
c:\tmp in the root - but you may not if you are running in a secured
windows environment.

So fix the launcher to only require powershell v3 or higher and don't
hardocde the temp directory.