[JENKINS-33520] Do not hard code the default ip to localhost.
The default ip for docker port mapping should not be 127.0.0.1 as this is
doomed to fail for containers created on a remote docker host.

Some containers need to be local as they are not fully treated as
containers but light weight virtual machines (e.g. FTP with PASV mode) so
still allow containers to be created with a specified address but
deprecate it.
Extended WithDocker so a test can be skipped if it really does need a
local docker instance.

Added the ability for a DockerContainer to customize its dockerfile at
tuntime (which was going to be used for FTP - but didn't work). Still as I
wrote it it may be usefull for someone so may as well be here...