[JENKINS-33520] Fix the unit test on windows.
The checking of the artifact.open() was effectivly a no-op.
On Firefox this causes popups to apear to verify the opening of the jar -
but more importantly a failure would do nothing - and this can be shows as
the loop to create the files is 0->7000 yet the loop would try and open
file 7042 which would fail but the test did not care.