[JENKINS-33520] Fix DeployPluginTest
Rather than write another batch/powershell script for this I tried
something different.  If running on windows then set Jenkins up with the
cygwin interpretter.   Works fine here - but obviously will fail for a
remote jenkins where the platform it is on is different to what the ATH is
running on - but that was going to be the case anyway.

Seems like we should have a utility method to ask Jenkins when it starts
so the tests can query and then do the right thing if they need.  (let a
Linux user run the ATH against tests on a windows machine for instance
would be nice!)