Upgrade to Selenium 2.53.0
htmlunit-driver is split from selenium

Upgrade to 2.53 from 2.52 as when getting an element it would scroll it to
the top of the page, however it would not ensure it was visible.
As such some elements would be hidden under the banner which would cause a
"click()" to fail as a different element would have got the event.