Fix the uninstall test
The uninstall input being a form submit does not have a path attribute so
the scroller workaround just ingnores it.
This causes the submit button to be hidden behind the banner which in turn
causes selenium to thrown an error when the button is clicked.

So given we have the form - just submit it...