[FIXED JENKINS-36045] Bump sleep for waiting for cidfile
The half-second sleep seems to hit race conditions periodically,
especially if the host is under load. Moving it to a second got rid of
the issues in my testing - from seeing it error out due to an empty
cidfile roughly half the time in 10 tests to zero times in 10 tests.