Switches the runDockerTests profile to use a JUnit Category.
The time spent instantiating each test and launching Firefox, etc when
you're just trying to run the Docker tests is...impressively
large. That's because the FilterRule doesn't kick in until after the
WebDriver has been injected and therefore the browser has been
launched. That also results in a lot of skipped tests rather than
those tests just not showing up at all in the Maven output, which IMO
is preferable.