[FIXED JENKINS-36125] Expand all Advanced when testing custom workspace
If we only click the first one, we're vulnerable if there's more than
one on the configure page, as could be the case depending on plugins
installed. To be safe, let's click all of the visible Advanced buttons
- we're not going to break anything by doing so, just increase visibility.

I tried to figure out a way to explicitly find the Advanced Project
Options advanced button, but I don't know xpath well enough to ensure
I get the right sibling of the <tr> that contains the Advanced Project
Options string, so...brute force!