[JENKINS-36045] Major change to how we get cids and docker logs
More testing needed to verify this actually does the trick
consistently, but it does *work* in the non-failure case for
sure. We're now detaching the container, getting the container ID from
the "docker run" output (since that's the output from a detached run)
and running a side process piping "docker logs --follow (cid)" to the
log file. That side process gets automatically killed at the end of
the test when we kill and rm the container as well.