Make minimum browser window size configurable.
Currently via environment variables.

We've had a lot of problems with scrolling to check
matrix/credentials/etc fields, which are...annoying. And fairly
pointless to work on, frankly. I'd strongly advocate for instead
allowing configurability of window geometry - I'd've thought that the
maximize call would do the trick on its own with a larger Xvnc
geometry, but it didn't, so I've made the browser height and width
configurable via environment variables.