Make minimum browser window size larger.
We've had a lot of problems with scrolling to check
matrix/credentials/etc fields, which are...annoying. And fairly
pointless to work on, frankly. I'd strongly advocate for instead
allowing configurability of window geometry - I'd've thought that the
maximize call would do the trick on its own with a larger Xvnc
geometry, but it didn't - in fact, it seems to do absolutely *nothing*
when run inside Xvnc.

So instead, let's just suck it up and bump the minimum window size to
an admittedly silly-large 1680x1050. Selenium does fine with detecting
visibility of parts of the window that are off the screen, so a
smaller screen than 1680x1050 causes no problems.

I'd be open to a smaller width too, but definitely bigger than
1280x960, the previous minimum.