Pulled column and portlet tests up to AbstractAnalysisTest.
Now all static analysis plug-ins have a common test that checks that the number of warnings is shown as a link in the warnings column and warnings table portlet.
Currently a test does not check the URL itself it rather clicks the link to see if the navigation works. This could be used as a basis for [JENKINS-39947].