login does not wait for the user to be logged in
If you have a large delay on your Authentication then login will return
but the login process will not have completed.

The upshot is you must never use WebElement.click() for anything that
will be gauranteed to change the current page (URL)
https://blog.codeship.com/get-selenium-to-wait-for-page-load/

Whilst this is a form I tried changing click() to submit() but this
caused a worse issue and did not even work.  So for now add sime methods
that will wait until the page has changed (ie the current element will
be stale.  This will occur even if the same page is returned)