Do not install form-element-path for tests annotated with @WithInstallWizard
Early installation of form-element-path causes to install implied dependencies
- in version that was detached. This can easily cause such plugin need to be updated
for other plugins to install correctly. While Plugin Manager is ready for this and
handle it well, Setup Wizard is reporting an issue and demand Jenkins to be restarted.

Note its page objects _can_ be made more robust to restart if needed, but since
it would happen every time, it would just obscure real issues this can detect. Not
to mention this is running Setup Wizard out of its expected mode of execution, as
it does not expect any plugins to be present effectively testing corner case rather
that main code paths.

Alternative approach used in the past was bumping form-element-path every time
plugin gets detached from core so it never have any implied dependencies to cause
problems. But the downside is someone has to update that plugin every time there
is a plugin detached from core, and it prevents the plugin load for older cores
making older cores essentially untestable with ATH!

This is essentially taking a trade-off of never using form-element-path for wizard
tests over getting them broken every so often _and_ restricting ATH usage to quite
recent cores.