Move Assertion classes from AssertJ to compile scope.
In normal projects, assertions are part of the target/test-classes folder. However this
approach does not work for ATH. In order to use these classes in plugins we need to
have them in compile scope. So after compilation these classes are now moved to target/classes.
So they will be automatically packaged into the ATH jar.