Better updated status
Previously the stale check was not working properly, if the
check failed with something different than `StaleElementException` it
 was swallowed by the outer catch and `true` was returned
making the callers believe the UC finished updating when it was not.

Not apparent in ci.jenkins.io but in other testing infra with custom wars
using more than one update center for example.