[JENKINS-63159] workaround tooltip preventing the click
whilst I have not seen this on the CI it was failing 100% of the time
locally with a test that did
```
 MatrixAuthorizationStrategy auth =
     security.useAuthorizationStrategy(MatrixAuthorizationStrategy.class);
 auth.addUser("alice").admin();
 auth.addUser("bob").developer();
```

the user bob was not added and the test timed out waiting for the alter
box.
it appears as though `click` would not work as the button was partially
obscured by the tooltip for the "administrator" permission that was just
clicked.

using keyboard navigation (space) to click the button instead worked for
me.

I was running selenium/firefox with the following commands:
`docker run -d -p 4444:4444 -p 5900:5900 -e no_proxy=localhost -e SCREEN_WIDTH=1680 -e SCREEN_HEIGHT=1090 selenium/standalone-firefox-debug:2.53.1`