Fix racecondition when saving forms
when saving the global configuration if you are going from an unsecured
to secured instance then you will have 2 page reloads

the first will be the action of saving causing a navigate to /manage on
completion.  however this may now be access controlled, so rather than
the normal manage page you may get the jenkins 403 page which is a very
plain HTML page with a meta refresh and some javascript of a similar
line that will then cause the page to navigate to the login page.

so when cehcking that "the page expects a form submission" text there is
a race condition as that is performed in 2 parts.
1) getting the root element (by.css("html")
2) getting the text of that element

however if you have just secured Jenkins in the manage page then you
could get the html element of the redirect page before the javascript
kicks in, and then when getting the text of that the page has reloaded
so the element is stale.

What makes this more ugly is as we are using css matching the error is
unhelpful
`TypeError: a in null`
changing the by.css to an xpath getting "/" showed a
staleElementException which led to this discovery.

so generially `save` no longer checks if the page expects a form
submission (not sure what that was expecting to catch)

and the security page object has some extra code so that we wait until
any redirect due to 403 has been handled.