wait for the scritp console
The script console is added by Javascript and not completely ready with
the page load.
Therefore we need to wait for the parts we want to manipulate to be
ready.

Fixes a flaky test observed when using master.runStript() with the following failure:

```
org.openqa.selenium.NoSuchElementException:
Unable to locate element: //*[@path='/script']

```