fix observed flakyness with click when adding publishers to a build
on some systems clicking a button will fail as the button is obscured by
the "Save"/"Apply" buttons in configuration.

whilst Selenium scrolls it appears as though the click is done whilst
the scrolling is still in progress (the diagnostic screenshot shows
everything is ok) but this is taken very shortly after the test failure.

With this just try and click a few times if the button is obscured.

Ideally we would re-write the Jenkins UX so that the Save/Apply buttons
did not float above other parts of the form but rather where at the
bottom and the form contents would stop above the buttons.
but that is not such an easy fix so for now just try a few times,