fix flakyness in obtaining the yui menus for dropdown
on some systems the menu would fail to be displayed when we tried to
click on it as it is shown asynchronously.

with this change we wait for it to be visible, and once clicked we wait
for it to be not visible also so we know the click has done something.