change the way we find and wait for menus to be displayed
the previous change (although it worked) probably worked by fluke.

repeatidly getting the element in a waitFor did not change anything
apart from taking a bit longer to do its thing (which solved the case
but for the wrong reasons).

the menu can be found regardless of if it is visible or not. so we just
find it and then wait for it to become visible.