Reduce flakiness on plugins check
I've observed this failing with `StaleElementReferenceException` from
time to time. I tracked it down to the plugin manager UI dynamic
behaviour (DOM elements show up and down) so the initially select
element becomes stale after a few milliseconds.

There is no way to know if the element will become stale or not, so the
only way I found to make this less flaky is by waiting 1 second and
retrying, then give up if it fails again.