fix subversion test.
The subversion container could no longer be built as the ubuntu version
was no longer supported and its repos removed.

Switched to Debain (from ubuntu) as this contained packages for
everything that was required.

Using ViewVC not webSVN now - but nothing was actually using this (we
could remove it entierly if we want.

took the opertunity to cleanup the docker file