Disable the MitM proxy
the MitM proxy spews copious amounts of logs at run time about
unknown_ca
Additionally the broser needs to be correctly configured to support the
MitM proxy by installing some CA certs.  DOing this on a local instance
is bad (its likely going to be re-used) and there is no API in selenium
to do this for browsers like remote-firefox and we do not customize the
contianer image, so firefox would not accept the response either.

So as it is of little use without hand tweaking the environemnt it runs
in and this is not done on CI that I know of just disable MitM.

if we do find that is really is required somewhere we can make it can be enabled by some property.