milestone only aborts previousl builds who are in *exactly* the previous ordinal
THis means we may not abort oler builds, but those builds themselves
should have been aborted.
So this becomes only an issue if the pipeline has started but not hit
the milestone (which is unlikely).  We could extend this in a limited
for loop to the previous N builds, but YAGNI.