Test against the LTS as well as the weekly
As we need to run the ATH against LTS version when they are released we
should ensure that the ATH itself has no regressions when run against
it.

This allows us to actually identify issues in the LTS as opposed to the
ATH no longer being compatable with it.
This requires that the ATH be compatable with the weekly and the LTS
line, which should have been mostly done but was never tested in the PR
builder.