if no proxy host has been set use the loopback address
previous to this change null would be passed in and then the browser would be configured with a proxy address of 0.0.0.0 which is complete garbage for a destination address.

This would work on some OS/kernel combinations and not others.
Given the document already states that it binds to localhost then this
makes it actually work as documented.