more deflaking SshSlavesPluginTest
the mouse position was enough that the credential store would receive the
focus when the "add" button was clicked, this caused the title to be
removed due to container-min.js `onCOntextMouseOver`
https://github.com/jenkinsci/jenkins/blob/master/war/src/main/webapp/scripts/yui/container/container-debug.js#L5356-L5359

as the tooltip is present regardless of the mouse state/location switch
to that instead.

I am not conviced with changing the title of an element direcctly like
it does - but this is very old behaviour and something to revisit when
we finally remove yahooUI from Jenkins core

Additionally the click on the credential submit may not be instant - so we wait until the click
has caused the submit button to be hidden before returing.