ByPass the FileDetector for things that are not files
The FileDetector replaces any string that is also a local file by a
temp string and uploads that file to the remote system.

This is generally a good thing, but sometimes the file is really a
string and not something local or that needs replaces.

This is always the case for the JVM path when used in an external system