use renovate to update the xxx-container browser modes
in order to the FallbackCOnfig to be matched we need to replace the
default ignorePaths as they ignore all test files by using **/test/**
however the ATH uses test in its production package name.

we should be updating test resources for containers where we are testing
plugins as incase of an upstream change in the target system we should
fail our plugin test, most of these are already covered by dependabot -
so they are excluded here.