Switch from mock-security-realm to JenkinsDatabaseRealm
mock-security-realm is not a production realm and we can do everything
we need with JenkinsDatabaseRealm

Whilst this requires a few more clicks, it requires one less plugin to
be installed.  Additionally if the mock-update-center is not in use this
can remove the need to hit the update center for this plugin.