Migrate from mock security realm (#1221)
* Allow signup without the mailer plugin

if the mailer plugins was not installed, the utility methods to signup
would not work and would force all callers to check conditionally and
use the long form methods.

this changes the simple signup(user) to work irrespective of it mailer
is installed or not.

* Switch from mock-security-realm to JenkinsDatabaseRealm

mock-security-realm is not a production realm and we can do everything
we need with JenkinsDatabaseRealm

Whilst this requires a few more clicks, it requires one less plugin to
be installed.  Additionally if the mock-update-center is not in use this
can remove the need to hit the update center for this plugin.