disable surefire reuse of forks
I have seen some interesting interactions and there is some static code
in the test-harness, so do not attempt to reuse surefire.

this should reduce a bit of flakyness and should not be noticable given
the overhead of starting a Jenkins insance for test.

This is also the setting we use for plugins in the plugin pom where the
%age overhead would be more noticable than here