remove IE and HTMLUnit support
IE is not supported by Jenkins so bringin in the dependency for testing
is is sub par.

Additionally HTMLUnit is not supported here, we are testing end-to-end
with real browsers.  anything HTMLUnit could be done with
RealJenkinsRule and HTMLUnit instead.

Add commons-net dependency that was previosly transitive but used by the
FTPFixture