Tidy up the contract around the HAR being recorded and the proxy being created
The proxy was created on the first call to getProxy and every call after
when the har had been recorded for the previous test.

the Fallback config was responsible for updating the proxy with the test
name

This spread the responsibility for the HAR recording between FallbackConfig
and the HarRecorder.

This change moves the initialisation of the har from the FallbackConfig to
the HarRecorder itself, and no longer shuts down the proxy if a test causes
the har to be recorded.

Whilst the Proxy would not be closed, this is the status quo today and it is
terminated when the surefire JVM exists.
So now we will start the proxy only once per surefire JVM.

This could be a candidate for auto-wiring but that is left for a future
change