replace dumpling with a call to jstack
with the current java11 I have not had a successful stacktrace when a
test timesout.

As we do not require parsing of the stacktrace we can replace dumpling
with standard calls.

jstack works on windows (without the jstack.exe).

Whilst this requires jstack to be on the path - as we are calling java
to launch this should be an acceptable requirement for the ATH.

fixes #1222