Installing plugins was flaky
it has been observed that plugin installation was flaky as the
installation clicked the button but then immediatly looked like there
where no plugins to update.

the HAR showed that the response to post the install had not finished
and then the call to the UpdateCenter was made and as the previous call
had not completed, there where no installation jobs present and so
deeemed that there was nothing to wait for.  the check then fails as the
plugin has not infact been installed

```
java.lang.IllegalArgumentException: workflow-multibranch plugin not installed
	at org.jenkinsci.test.acceptance.po.Plugin.<init>(Plugin.java:49)
	at org.jenkinsci.test.acceptance.po.Jenkins.getPlugin(Jenkins.java:214)
	at org.jenkinsci.test.acceptance.po.UpdateCenter.waitForInstallationToComplete(UpdateCenter.java:98)
	at org.jenkinsci.test.acceptance.po.PluginManager.installPlugins(PluginManager.java:259)
	at org.jenkinsci.test.acceptance.junit.WithPlugins$RuleImpl$1.installPlugins(WithPlugins.java:191)
```