Fix flakiness on Gitlab tests (First PR, Local Attempt 1652756)
I detected most of the times flakiness was caused by a lack of resources in the Gitlab
container, as a result many API calls at random returned a 502 error code
and an HTML page asking for the user to refresh...

I have added a custom configuration to the Gitlab fixture which should help in
resource constrained environments and increased the request timeout. Also
not it uses the readiness checks instead of html web scrapping.

Also some small code changes here and there like closing the gitlab4j clients
or slighly reduce the scope of some tests.