Add oic auth e2e (#1701)
* Add test for oic-auth plugin

* Retrieve and check the roles from the OICD provider

* Update src/main/java/org/jenkinsci/test/acceptance/po/OicAuthSecurityRealm.java

Co-authored-by: James Nord <jtnord@users.noreply.github.com>

* Apply suggestions

* Add the WhoAmI page

* Fix permissions on the docker socket

TestContainers does not use docker, but talks directly to the docker
socket.  The permissions on this socket come from the host where it is
mapped and the docker groupid may not match what we have in the
container.
So allow th arg to be passed through at build time and add the ath-user
to the docker group so it has the permissions.

We retain the legacy suid on the docker binary as we publish the
container and there is only a single test so far using this
test-containers.  (this can be revistied if required).

* Incremental build

* Remove unused libraries

* Add missing dependencies

* New incremental

* New incremental

* update to 2.476

* document testcontainers requirement

* update keycloak container vie renovate

* s/TestContainers/Testcontainers

Testcontainers is with a lower case c

---------

Co-authored-by: James Nord <jtnord@users.noreply.github.com>
Co-authored-by: Basil Crow <me@basilcrow.com>