Remove JabberPluginTest that has been ignored for over 10 years
Test was introduced in Jun 2014 with commit:

* 1cc86a3c3a0f7088f2ace8d5cabe751831672b93

Test was first ignored in Jun 2014 with commit:

* 10a5ab6745f46b17dba066f152004fb3f4fef9c1

That commit was merged Jun 30, 2014 as part of:

* f225f2e39546ef67dde3ce9b95d85b67b40dc5f5

The contributing guide[1] says that after 90 days, a test can be removed if:

> Test is determined fragile/broken when it is failing despite no defects
> are present in tested component.
>
> * JIRA issue will be created notifying author of the test or plugin authors
> * Test would be @Ignored while the JIRA is open
> * At the end of the time period, the test would be removed from ATH.
>
> The situation is expected to be handled in 90 days since reported. In
> case of no corrective action, the test will be removed from ATH.

I confirmed that `mvn clean -DskipTests verify` passes with the tests
deleted.  Will rely on ci.jenkins.io to do a deeper set of tests to
confirm no harm from the removal.

[1] https://github.com/jenkinsci/acceptance-test-harness/blob/master/docs/CONTRIBUTING.md#dealing-with-fragilebroken-tests)