Test revert of updateComputerList performance improvement
Reverting pull request is:

* https://github.com/jenkinsci/jenkins/pull/10537

Original pull request is:

* https://github.com/jenkinsci/jenkins/pull/10494

Acceptance test harness creates an agent using command launcher and
saves the agent definition.  Before the change, that saved definition
would immediately start the newly defined agent.  After the change,
the agent does not start and the inbound agent list of example commands
is displayed.  It seems as though the new agent does not fully understand
that it is a command launcher agent.