Wait for the notification to disappear when adding credentials (#2145)
Wait for the notifcation to disappear when adding credentials

WHen adding a credential from the dialog the notifaction bar can pop up
and obscure some elements.  Ensure that the notification has been
displayed and is currently hidden before returning from `add`

Update SshCredential to extend from BaseStandardCredentials and remove
the custom implementation of add.