Increase timeout on ant plugin tests (#2170)
increase build timeout

The test was timing out as the ant build took longer than 2 minutes (the
default timeout for waiting for a build to complete in shouldSucceed

We now wait for 4 minutes.  Whilst we are still using external resources
(so this will always have the potential to be flaky, this is actually
testing the functionality that Jenkins does download from these external
resources).

Alternative would be to change the production code so we could overwrite
tool installer URLs with some other URL (but then we would not fail if
the upstream server changed layout).  This is the just good enough quick
fix