disable implicit waiting
implicit waiting is a sure fire sign that the tests or page objects are
flaky.  If something needs to wait for something async it should do it in a `waitFor`
Doing it blindly leads to cases where it looks like the code is fin but
is only working by chance on the current hardware and can fail elsewhere
on different hardware with different timings.