only check CSP if test passes
if a test failes do not attempt to check for CSP violations.
Doing the check prevents the screenshot capture from capturing the
screenshot of the page we where on when the test failed as the rule does
navigation to obtain the results of the check.

The check could also be improved in the plugin to allow an API like
access so that we do not even need to use Selenium but this change should
be good enough.