Bump to keycloak 26.4.2 and adapt tests
Since Keycloak 26.1.0 when requesting an offline_scope the online
session will be revoked if it has not been used before.
So in essence when you authenticate to KeyCloak as part of the Jenkins
OIDC auth dance the session to keycloak will be revoked so you will no
longer be logged in on keycloak.