simplify Dialog handling
A dialog is not an alert and attempting to fallback will run the
runnable twice which is a silly thing to do.

Update so that we wait for the dialog to be shown and then use the
default button for progressing.

In the case of input find the input within the dialog to reduce the
search scope.